tests/requirements.txt
google-adk>=1.0.0
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
whois>=0.9.27
typing-extensions>=4.5.0
pydantic>=2.0.0
